# Deployment Controller - avoid unnecessary copying of pod objects in getPodMapForDeployment