--- a/pkg/controller/deployment/deployment_controller.go
+++ b/pkg/controller/deployment/deployment_controller.go
@@ -366,7 +366,7 @@ func (dc *DeploymentController) deletePod(obj interface{}) {
 		}
 		numPods := 0
 		for _, podList := range podMap {
-			numPods += len(podList.Items)
+			numPods += len(podList)
 		}
 		if numPods == 0 {
 			dc.enqueueDeployment(d)
@@ -525,7 +525,9 @@ func (dc *DeploymentController) getReplicaSetsForDeployment(d *apps.Deployment) 
 //
 // It returns a map from ReplicaSet UID to a list of Pods controlled by that RS,
 // according to the Pod's ControllerRef.
-func (dc *DeploymentController) getPodMapForDeployment(d *apps.Deployment, rsList []*apps.ReplicaSet) (map[types.UID]*v1.PodList, error) {
+// NOTE: The pod pointers returned by this method point the the pod objects in the cache and thus
+// shouldn't be modified in any way.
+func (dc *DeploymentController) getPodMapForDeployment(d *apps.Deployment, rsList []*apps.ReplicaSet) (map[types.UID][]*v1.Pod, error) {
 	// Get all Pods that potentially belong to this Deployment.
 	selector, err := metav1.LabelSelectorAsSelector(d.Spec.Selector)
 	if err != nil {
@@ -536,9 +538,9 @@ func (dc *DeploymentController) getPodMapForDeployment(d *apps.Deployment, rsLis
 		return nil, err
 	}
 	// Group Pods by their controller (if it's in rsList).
-	podMap := make(map[types.UID]*v1.PodList, len(rsList))
+	podMap := make(map[types.UID][]*v1.Pod, len(rsList))
 	for _, rs := range rsList {
-		podMap[rs.UID] = &v1.PodList{}
+		podMap[rs.UID] = []*v1.Pod{}
 	}
 	for _, pod := range pods {
 		// Do not ignore inactive Pods because Recreate Deployments need to verify that no
@@ -548,8 +550,8 @@ func (dc *DeploymentController) getPodMapForDeployment(d *apps.Deployment, rsLis
 			continue
 		}
 		// Only append if we care about this UID.
-		if podList, ok := podMap[controllerRef.UID]; ok {
-			podList.Items = append(podList.Items, *pod)
+		if _, ok := podMap[controllerRef.UID]; ok {
+			podMap[controllerRef.UID] = append(podMap[controllerRef.UID], pod)
 		}
 	}
 	return podMap, nil