**What type of PR is this?**
/kind feature

**What this PR does / why we need it**:
It improves DeploymentController memory/cpu consumption by avoid unnecessary copying of pod objects.
As the benchmark shows it speeds up the method~x4 and reduces memory consumption ~x20.

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkGetPodMapForDeployment-12     276121        72591         -73.71%

benchmark                              old allocs     new allocs     delta
BenchmarkGetPodMapForDeployment-12     241            238            -1.24%

benchmark                              old bytes     new bytes     delta
BenchmarkGetPodMapForDeployment-12     554025        28956         -94.77%
```

I've also run a gce-100 perf test and the changes seems to be visibly improving memory/cpu consumption of the kube-controller-manager (see https://github.com/kubernetes/kubernetes/issues/79898)
![3EObupeKZLQ](https://user-images.githubusercontent.com/2604887/60898721-8baac080-a269-11e9-87a3-d94a805c88ba.png)

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```

/sig scalability
/priority important-soon